+++ packages/perseus-core/src/types.ts
@@ -1,0 +1,22 @@
+// Types that can be shared between Perseus packages
+// ideally without causing circular dependencies
+
+type State = any;
+
+export interface RendererInterface {
+    getSerializedState(): State;
+    restoreSerializedState(state: State, callback?: () => void): void;
+    scoreInput(): KEScore;
+    blur(): void;
+    focus(): boolean | null | undefined;
+    props: any;
+}
+
+export type KEScore = {
+    empty: boolean;
+    correct: boolean;
+    message?: string | null | undefined;
+    suppressAlmostThere?: boolean | null | undefined;
+    guess: any;
+    state: any;
+};